requirements.txt
# Core dependencies
psutil>=5.9.0
flask>=3.0.0
plyer>=2.1.0

# Database
sqlalchemy>=2.0.0

# WebSocket support for real-time updates
flask-socketio>=5.3.0
python-socketio>=5.10.0

# System tray integration
pystray>=0.19.0
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0

# Optional: Email notifications (using built-in smtplib)
# Optional: SMS notifications
twilio>=8.10.0

# Scheduling (using built-in sched and threading)
# Windows Task Scheduler (using subprocess)